Move context.js to components/common
It exports a component, and is only used by components.
